chore: update playwright and add auto update
