+++ .github/workflows/02-e2e.yml
@@ -12,7 +12,7 @@ jobs:
     name: 🧪🎭 - ${{ matrix.framework }}:${{ matrix.shard }}
     runs-on: ubuntu-latest
     container:
-      image: mcr.microsoft.com/playwright:v1.40.0
+      image: mcr.microsoft.com/playwright:v1.40.1
     strategy:
       fail-fast: false
       matrix:
