+++ e2e/Dockerfile
@@ -1,4 +1,4 @@
-FROM mcr.microsoft.com/playwright:v1.40.0 as playwright
+FROM mcr.microsoft.com/playwright:v1.40.1 as playwright
 
 COPY e2e /e2e
 WORKDIR /e2e
